#ECOAPI-830 - Transformando a classe do Authentication de testes unitários em classes de exemplo

+++ src/main/java/br/com/senior/core/authentication/pojos/LoginMFAInput.java
@@ -22,6 +22,6 @@ public class LoginMFAInput {
     /**
      * Código de validação gerado pelo Google Authenticator
      */
-    Long validationCode;
+    int validationCode;
 
 }
